+++ src/Data/Semigroup/Abelian.hs
@@ -5,7 +5,6 @@ module Data.Semigroup.Abelian
     , fromNonEmpty
     ) where
 
-import           Data.Constraint (Dict (..))
 import           Data.IntSet (IntSet)
 import           Data.List.NonEmpty (NonEmpty)
 import qualified Data.List.NonEmpty as NE
@@ -29,7 +28,7 @@ import           Data.Algebra.Free
     ( AlgebraType
     , AlgebraType0
     , FreeAlgebra (..)
-    , Proof (..)
+    , proof
     )
 
 -- |
@@ -101,5 +100,5 @@ instance FreeAlgebra FreeAbelianSemigroup where
         toNonEmpty_ :: Map a Integer -> NonEmpty a
         toNonEmpty_ = NE.fromList . concat . map (uncurry replicate_) . Map.toList
 
-    proof  = Proof Dict
-    forget = Proof Dict
+    codom  = proof
+    forget = proof